Algoflow_python_agent/pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents==1.5.1",
    "livekit-plugins-deepgram==1.5.1",
    "livekit-plugins-groq==1.5.1",
    "livekit-plugins-silero==1.5.1",
    "livekit-plugins-sarvam==1.5.1",
    "livekit-plugins-openai==1.5.1",
    "python-dotenv==1.2.2",
    "numpy==2.4.3",
    "aiohttp==3.13.3",
    "selenium==4.27.1",
    "beautifulsoup4==4.13.4",
    "webdriver-manager==4.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
